Move bulk upload into model (BL-11581)
