+++ src/BloomExe/Publish/BloomLibrary/BloomLibraryUploadControl.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.IO;
@@ -9,11 +8,9 @@ using System.Linq;
 using System.Windows.Forms;
 using Bloom.Book;
 using Bloom.Collection;
-using Bloom.ErrorReporter;
 using Bloom.ImageProcessing;
 using Bloom.MiscUI;
 using Bloom.TeamCollection;
-using Bloom.Utils;
 using Bloom.web;
 using Bloom.web.controllers;
 using Bloom.WebLibraryIntegration;
@@ -879,109 +876,8 @@ namespace Bloom.Publish.BloomLibrary
 
 		private void BulkUpload(string rootFolderPath)
 		{
-			var target = BookUpload.UseSandbox ? UploadDestination.Development : UploadDestination.Production;
-
-			var bloom = Application.ExecutablePath;
-			var command = $"\"{bloom}\" upload \"{rootFolderPath}\" -u {_userId.Text} -d {target}";
-			if (SIL.PlatformUtilities.Platform.IsLinux)
-				command = $"/opt/mono5-sil/bin/mono {command}";
-
-			ProcessStartInfo startInfo;
-			if (SIL.PlatformUtilities.Platform.IsWindows)
-			{
-				startInfo = new ProcessStartInfo()
-				{
-					FileName = "cmd.exe",
-					Arguments = $"/k {MiscUtils.EscapeForCmd(command)}",
-					
-					WorkingDirectory = Path.GetDirectoryName(Application.ExecutablePath)
-				};
-			}
-			else
-			{
-				string program = GetLinuxTerminalProgramAndAdjustCommand(ref command);
-				if (String.IsNullOrEmpty(program))
-				{
-					_progressBox.Clear();
-					_progressBox.WriteMessage("Cannot bulk upload because unable to find terminal window for output messages.");
-					return;
-				}
-				startInfo = new ProcessStartInfo()
-				{
-					FileName = program,
-					Arguments = command,
-					WorkingDirectory = Path.GetDirectoryName(Application.ExecutablePath)
-				};
-				// LD_PRELOAD is a Linux environment variable for a shared library that should be loaded before any other library is
-				// loaded by a program that is starting up.  It is rarely needed, but the mozilla code used by Geckofx is one place
-				// where this feature is used, specifically to load a xulrunner patch (libgeckofix.so) that must be in place before
-				// xulrunner can be initialized for GeckoFx60 on Linux.  This must be in place in the environment before launching
-				// any process (such as Bloom, here) that will initialize xulrunner, but may cause problems for other programs so
-				// it is best not to have it in the environment unless we know it is needed.  In particular having LD_PRELOAD set to
-				// load libgeckofix.so is known to cause problems when running some programs (possibly only BloomPdfMaker.exe) using
-				// CommandLineRunner.  To guard against this Program.Main() removes it from the environment, but here we need to
-				// temporarily restore it so it can be inherited by the instance of Bloom we are about to launch. Fortunately, it's
-				// easy to reconstruct.
-				var xulRunner = Environment.GetEnvironmentVariable("XULRUNNER");
-				if (!String.IsNullOrEmpty("xulRunner"))
-					Environment.SetEnvironmentVariable("LD_PRELOAD", $"{xulRunner}/libgeckofix.so");
-			}
-
-			Process.Start(startInfo);
 			_progressBox.Clear();
-			_progressBox.WriteMessage("Starting bulk upload in a terminal window...");
-			_progressBox.WriteMessage("This process will skip books if it can tell that nothing has changed since the last bulk upload.");
-			_progressBox.WriteMessage("When the upload is complete, there will be a file named 'BloomBulkUploadLog.txt' in your collection folder.");
-			var url = $"{BloomLibraryUrls.BloomLibraryUrlPrefix}/{_model.Book.CollectionSettings.DefaultBookshelf}";
-			_progressBox.WriteMessage("Your books will show up at {0}", url);
-			if (SIL.PlatformUtilities.Platform.IsLinux)	// LD_PRELOAD interferes with CommandLineRunner and GeckoFx60 on Linux
-				Environment.SetEnvironmentVariable("LD_PRELOAD", null);
-		}
-
-		private void _uploadSource_SelectedIndexChanged(object sender, EventArgs e)
-		{
-			UpdateDisplay();
-		}
-
-		private string QuoteQuotes(string command)
-		{
-			return command.Replace("\\", "\\\\").Replace("\"", "\\\"");
-		}
-
-		private string GetLinuxTerminalProgramAndAdjustCommand(ref string command)
-		{
-			// See https://askubuntu.com/questions/484993/run-command-on-anothernew-terminal-window
-
-			if (RobustFile.Exists("/usr/bin/gnome-terminal"))	// standard for GNOME (Ubuntu/Wasta)
-			{
-				// /usr/bin/gnome-terminal -- /bin/bash -c "bloom upload \"folder\" -u user -d dest; read line"
-				command = $"-- /bin/bash -c \"{QuoteQuotes(command)}; read line\"";
-				return "/usr/bin/gnome-terminal";
-			}
-			if (RobustFile.Exists("/usr/bin/terminator")) // popular alternative
-			{
-				// /usr/bin/terminator -x /bin/bash -c "bloom upload \"folder\" -u user -d dest; read line"
-				command = $"-x /bin/bash -c \"{QuoteQuotes(command)}; read line\"";
-				return "/usr/bin/terminator";
-			}
-			if (RobustFile.Exists("/usr/bin/xfce4-terminal"))    // standard for XFCE4 (XUbuntu)
-			{
-				// /usr/bin/xterm -hold -x /bin/bash -c "bloom upload \"folder\" -u user -d dest"
-				command = $"-T \"Bloom upload\" --hold -x /bin/bash -c \"{QuoteQuotes(command)}\"";
-				return "/usr/bin/xfce4-terminal";
-			}
-			if (RobustFile.Exists("/usr/bin/xterm"))	// antique original (slightly better than nothing)
-			{
-				// /usr/bin/xterm -hold -x /bin/bash -c "bloom upload \"folder\" -u user -d dest"
-				command = $"-T \"Bloom upload\" -hold -e /bin/bash -c \"{QuoteQuotes(command)}\"";
-				return "/usr/bin/xterm";
-			}
-			// Neither konsole nor qterminal will launch with Bloom.  The ones above have been tested on Wasta 20.
-			// symbol lookup error: /usr/lib/x86_64-linux-gnu/qt5/plugins/styles/libqgtk2style.so: undefined symbol: gtk_combo_box_entry_new
-			// I suspect because they're still linking with GTK2 while Bloom has to use GTK3 with Geckofx60.
-
-			// Give up.
-			return null;
+			_model.BulkUpload(rootFolderPath, _progressBox);
 		}
 
 		private void _targetProduction_CheckedChanged(object sender, EventArgs e)
